rename folder 'js' to 'utils'
