+++ src/commands/profile/profile.ts
@@ -1,8 +1,8 @@
 import { Client, Message } from 'discord.js'
 import { Db } from 'mongodb'
 import * as ejs from 'ejs';
-import imGenerator from '../../js/img';
-import utilities from '../../js/utilities'
+import imGenerator from '../../utils/img';
+import utilities from '../../utils/utilities'
 let lastComboColor:number;
 
 module.exports.run = (bot:Client, msg:Message, args:string[], db:Db) => {
